tt/pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "tt"
version = "0.1.0"
description = "Translation tool - copies implementation from ghostfolio_pytx_example"
requires-python = ">=3.11"
dependencies = [
    "esprima2>=5.0.1",
    "fastapi>=0.115",
    "uvicorn[standard]>=0.30",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.2",
    "requests>=2.32",
]

[project.scripts]
tt = "tt.cli:main"

[tool.setuptools.packages]
find = {}
